streamlit/requirements.txt
# Web application development framework
pipenv
sqlalchemy
cryptography
numpy==1.19.2
seaborn==0.11.0
matplotlib==3.1.3
pandas==1.1.2
pymysql==0.10.1
streamlit==0.71.0
mysql-connector-python
